status update
fixed status update
